test/dbt_integration_tests/dbt-snowplow-web/requirements.txt
requests
snowflake-connector-python
pandas==2.2.3
plotly==5.24.1
kaleido==0.2.1
python-dotenv
